Add property change support and observer tests to Konto class
The changes include the addition of property change support for the Konto.java class, which facilitates observation of changes in properties like kontostand, gesperrt, and waehrung. This observer pattern implementation allows for a more efficient interaction with these properties. Moreover, unit tests have been added for this observer functionality, helping ensure the correct behavior of the software.